+++ makefiles/posix/config_posix_sitl.mk
@@ -72,9 +72,10 @@ MODULES		+= lib/geo_lookup
 MODULES		+= lib/conversion
 
 #
-# Linux port
+# POSIX port
 #
 MODULES		+= platforms/posix/px4_layer
+MODULES		+= platforms/posix/work_queue
 MODULES		+= platforms/posix/drivers/accelsim
 MODULES		+= platforms/posix/drivers/gyrosim
 MODULES		+= platforms/posix/drivers/adcsim
