+++ src/platforms/posix/work_queue/dq_rem.c
