+++ src/platforms/posix/work_queue/dq_remfirst.c
