+++ src/platforms/posix/work_queue/sq_remfirst.c
