+++ src/platforms/posix/work_queue/work_cancel.c
