+++ src/platforms/qurt/px4_layer/module.mk
@@ -42,25 +42,9 @@ SRCDIR=$(dir $(MODULE_MK))
 SRCS		 = 	\
 			px4_qurt_impl.cpp \
 			px4_qurt_tasks.cpp  \
-			hrt_thread.c \
-                        hrt_queue.c \
-                        hrt_work_cancel.c \
-			work_thread.c \
-			work_queue.c \
-			work_lock.c \
-			work_cancel.c \
 			lib_crc32.c \
 			drv_hrt.c \
-			queue.c \
-			dq_addlast.c \
-			dq_remfirst.c \
-			sq_addlast.c \
-			sq_remfirst.c \
-			sq_addafter.c \
-			dq_rem.c \
-                        hrt_work.c \
                         qurt_stubs.c \
-                        qurt_hacks.c \
                         main.cpp
 ifeq ($(CONFIG),qurt_hello)
 SRCS +=			commands_hello.c
